+++ yampt.editor/source/dialog/settings/merge_settings_view.hpp
@@ -1,7 +1,6 @@
 #pragma once
 
 #include <string>
-#include <unordered_map>
 #include <QWidget>
 
 class QCheckBox;
@@ -24,10 +23,9 @@ private:
 	void setup_fixes_tab();
 
 	QTabWidget * m_tabs = nullptr;
-	std::unordered_map<std::string, QCheckBox *> m_type_checkboxes;
+	QLineEdit * m_ignore_fields_edit = nullptr;
 	QLineEdit * m_exclusion_edit = nullptr;
 	QCheckBox * m_fog_fix_check = nullptr;
 	QCheckBox * m_summon_fix_check = nullptr;
 	QCheckBox * m_cell_name_fix_check = nullptr;
-	QLineEdit * m_ignore_sub_records_edit = nullptr;
 };